### 2016-08-12 #49 @whedon
[REVIEW]: TPIB - Transformation Peak Iterative Baseline
**Submitting author:** @drcassar (<a href="http://orcid.org/0000-0001-6472-2780">Daniel Cassar</a>)
**Repository:** <a href="https://github.com/drcassar/TPIB" target ="_blank">https://github.com/drcassar/TPIB</a>
**Version:** v1.0.0
