### 2016-08-12 ca91fbc
Update paper.md
