### 2016-08-16 #49 @pjotrp
Someone with a physics/chemical background can pick this up
