### 2016-09-22 #49 @kyleniemeyer
Hi @nicoguaro & @khinsen: would either of you be able to review this submission?
