### 2016-09-22 #49 @arfon
> @arfon ah, this one came before the new system, right? I'll just manually change the label once the review starts.

Yep. Still do the `@whedon assign @username as reviewer` though as this gives the reviewer permissions to update the checklist on the issue.
