### 2016-10-08 #49 @nicoguaro
@kyleniemeyer, I didn't notice that you asked me on this one.

This package consists of a single Python file. I'm not saying it because that's something wrong or right. But the question rises, is there a "minimum"? Or the scientific relevance of the software is what determine it?

It does not have installation instructions, tests, nor contribution guidelines. I think that the folder structure can be improved as well. A better statement of need would be good as well, since I suppose that it can be used in more general contexts than the example provided.
