### 2016-10-08 #49 @drcassar
Hello @nicoguaro, thanks for pointing out ways to improve the submission. You see, this is my first time uploading a code to github so I'm still not familiar with all the good practices. I'll look into other accepted  JOSS submissions and check how to improve mine following your suggestions.
The statement of need is focused on differential thermoanalysis measurements. It's one of the most used characterization techniques in the field of materials science. I too suppose it may be useful in other contexts, but I would like to stress that this is not a small field and the reproductibility problem mentioned in the statement (due to the closed nature of various commercial software) is an actual problem.
Please, let me know if it is not OK for me to be posting here within review process.
