### 2016-10-08 #49 @kyleniemeyer
Hi @drcassar, this is a perfectly fine place for this discussion. I agree that you should follow @nicoguaro's suggestions about structuring the repo (especially such that it can be installed/distributed as a package); you can look at other Python packages published in JOSS as examples, or use these guidelines: http://docs.python-guide.org/en/latest/writing/structure/

In addition, there are some suggestions about tests, contribution guidelines, and explanation of how to run the examples. I would work towards making these changes—please let us know when it's ready  to take another look.

@nicoguaro No problem, thanks for looking at it. We don't have a minimum requirement in terms of what is suitable; if the software has a research application and is feature-complete, then it should be acceptable. This package certainly seems focused on one particular technique, but that is fine.
