### 2016-10-08 #49 @nicoguaro
@kyleniemeyer, I agree. I just wanted to point it out, some of those points are still grey for me.

@drcassar, can the the software be used for some other measurements and not just differential thermoanalysis measurements? Let's say... some spectral techniques.
