### 2016-10-10 #49 @drcassar
Hi @kyleniemeyer, and thanks for the info. I'll read more and update my submission, letting you guys know when I finish the update. Right now I'm studying for a very tough exam, so I'll probably not be able to fix this too quickly. Let me know if this is a problem and I can perhaps remove the submission and try again later.
Good suggestion @nicoguaro, I'll discuss with some experts in this field to see if it could be useful.
Cheers
